fix generators array creation;
